requirements.txt
python-dateutil
slumber
astropy
pyvo
